# schemas/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
      "AuditEvent": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Hash"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "outcome": {
            "default": "attempt",
            "title": "Outcome",
            "type": "string"
          },
          "request_id": {
            "default": "",
            "title": "Request Id",
            "type": "string"
          },
          "subject_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Hash"
          }
        },
        "required": [
          "event_id",
          "occurred_at",
          "action"
        ],
        "title": "AuditEvent",
        "type": "object"
      },
...
      "CollectionResponse_AuditEvent_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/CursorPagination"
          }
        },
        "required": [
          "data",
          "pagination",
          "meta"
        ],
        "title": "CollectionResponse[AuditEvent]",
        "type": "object"
      },
...
      "CursorPagination": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "CursorPagination",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "INVALID_ARGUMENT",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "RATE_LIMITED",
          "SOURCE_UNAVAILABLE",
          "PROCESSING_FAILED",
          "INTEGRITY_FAILED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "IdentityResponse": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "subject",
          "kind",
          "roles",
          "scopes"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
...
      "ProvenanceRecord": {
        "additionalProperties": false,
        "properties": {
          "analysis_id": {
            "anyOf": [
              {
                "pattern": "^ana_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "input_ids": {
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            },
            "title": "Input Ids",
            "type": "array"
          },
          "operation": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Operation",
            "type": "string"
          },
          "provenance_id": {
            "pattern": "^prov_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Provenance Id",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "schema_version": {
            "const": "1",
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "subject_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "provenance_id",
          "subject_id",
          "input_ids",
          "operation",
          "recorded_at",
          "content_hash"
        ],
        "title": "ProvenanceRecord",
        "type": "object"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "SignedManifest": {
        "additionalProperties": false,
        "properties": {
          "chain_hashes": {
            "items": {
              "type": "string"
            },
            "title": "Chain Hashes",
            "type": "array"
          },
          "key_id": {
            "title": "Key Id",
            "type": "string"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecord"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Records",
            "type": "array"
          },
          "signature": {
            "title": "Signature",
            "type": "string"
          },
          "version": {
            "default": "aegis-provenance-1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "records",
          "chain_hashes",
          "key_id",
          "signature"
        ],
        "title": "SignedManifest",
        "type": "object"
      },
...
      "SingleResponse_IdentityResponse_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IdentityResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[IdentityResponse]",
        "type": "object"
      },
...
      "SingleResponse_VerificationResult_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VerificationResult"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[VerificationResult]",
        "type": "object"
      },
...
      "VerificationRequest": {
        "additionalProperties": false,
        "properties": {
          "contents": {
            "additionalProperties": {
              "type": "string"
            },
            "maxProperties": 100,
            "title": "Contents",
            "type": "object"
          },
          "manifest": {
            "$ref": "#/components/schemas/SignedManifest"
          }
        },
        "required": [
          "manifest",
          "contents"
        ],
        "title": "VerificationRequest",
        "type": "object"
      },
      "VerificationResult": {
        "additionalProperties": false,
        "properties": {
          "chain_valid": {
            "title": "Chain Valid",
            "type": "boolean"
          },
          "content_verified": {
            "title": "Content Verified",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "signature_valid": {
            "title": "Signature Valid",
            "type": "boolean"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "signature_valid",
          "chain_valid",
          "content_verified",
          "reason"
        ],
        "title": "VerificationResult",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Modular monolith with durable offline ingestion and canonical document retrieval.",
    "title": "AegisNews",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/security/audit": {
      "get": {
        "operationId": "audit_events_api_v1_security_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse_AuditEvent_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Events",
        "tags": [
          "security"
        ]
      }
    },
    "/api/v1/security/me": {
      "get": {
        "operationId": "me_api_v1_security_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_IdentityResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "security"
        ]
      }
    },
    "/api/v1/security/verify": {
      "post": {
        "operationId": "verify_api_v1_security_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_VerificationResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Verify",
        "tags": [
          "security"
        ]
      }
    },
...
  },
...
}